refactor(config): Split out assets
